+++ src/App.tsx
@@ -20,7 +20,7 @@ function App() {
       {/* Welcome */}
       <section id="welcome" className="grid">
         <div className="welcome-text">
-          <h1>Cristian</h1>
+          <h1>Cristiano</h1>
           <h4>Frontend Engineer junior</h4>
           <p className="leading">Hello, I am <span style={stylito}>Cristian Tinipuclla</span>, bachelor of Computer Science and Systems Engineering from USIL. Apart from Frontend knowledge, I have a B2 level of English. I am very communicative, proactive and I love challenges.</p>
           <a href="#portfolio" className="button">View my work</a>