+++ CHANGELOG.md
@@ -12,3 +12,7 @@ project adheres to [Semantic Versioning](https://semver.org).
 ## [1.0.1] - 2024-02-15
 
 - *gitignore*: .vite folder to be ignored.
+
+## [test1] - 2024-02-17
+## [test2] - 2024-02-17
+- *App.tsx* minimal change in h1